requirements.txt
django-crispy-forms
requests
djangorestframework-bulk
django-material 
django-viewflow
python-dev-tools
libpq-dev
psycopg2
django-q
# yum install python3.6-devel